Extract <AvatarUploader />
